# web/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisRequest": {
        "additionalProperties": false,
        "description": "Analyse one Strategy, as-of one moment.\n\nThere is no Strategy type: a Strategy is an ordered list of Legs and nothing more\n(CONTEXT.md). The order is the order the trader built them in, and it is preserved\nbecause the per-Leg Greeks table (#27) is read alongside the Legs on screen.\n\nThe moment is what the whole response is served as-of - the Chain, the Entry\nPremiums and the volatilities all come from it, so it is asked for once here rather\nthan per Leg.\n\n**The Expiry is the other way round** (#71). It is a property of each Leg and not of\nthe request, because a moment is one thing a Strategy is looked at from and an Expiry\nis one thing each contract in it has. Nothing here restates it, so nothing here can\ncontradict a Leg.\n\nA Strategy whose Legs span more than one Expiry is **refused** - see\n`strategy.sole_expiry`. Not refused here: the rule is about what can be drawn rather\nthan about what can be typed, and `strategy.MixedExpiry` says which two series were\nnamed where a validation envelope would say `body -> legs`.",
        "properties": {
          "legs": {
            "items": {
              "$ref": "#/components/schemas/LegRequest"
            },
            "title": "Legs",
            "type": "array"
          },
          "moment": {
            "title": "Moment",
            "type": "string"
          }
        },
        "required": [
          "moment",
          "legs"
        ],
        "title": "AnalysisRequest",
        "type": "object"
      },
      "AnalysisResponse": {
        "description": "Everything about one Strategy, in one response.\n\nDeliberately fat (#23). Splitting it would mean several round trips carrying the\nsame Legs and recomputing the same curve, and the trader would watch the numbers\narrive after the chart they belong to. #27 and #29 add the Greeks and the Payoff\nTable to this same response rather than to endpoints of their own.",
        "properties": {
          "curve": {
            "$ref": "#/components/schemas/Curve"
          },
          "discount": {
            "title": "Discount",
            "type": "number"
          },
          "forward": {
            "title": "Forward",
            "type": "number"
          },
          "greeks": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/LegGreeks"
            },
            "title": "Greeks",
            "type": "array"
          },
          "metrics": {
            "$ref": "#/components/schemas/Metrics"
          },
          "moment": {
            "title": "Moment",
            "type": "string"
          },
          "spot": {
            "title": "Spot",
            "type": "number"
          },
          "table": {
            "$ref": "#/components/schemas/Curve",
            "default": {
              "pnl_at_expiry": [],
              "spot": []
            }
          },
          "total_greeks": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LegGreeks"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "moment",
          "spot",
          "forward",
          "discount",
          "curve",
          "metrics"
        ],
        "title": "AnalysisResponse",
        "type": "object"
      },
...
      "ChainResponse": {
        "description": "The Chain a trader browses, as-of one moment.",
        "properties": {
          "discount": {
            "title": "Discount",
            "type": "number"
          },
          "expiry": {
            "title": "Expiry",
            "type": "string"
          },
          "forward": {
            "title": "Forward",
            "type": "number"
          },
          "forward_method": {
            "enum": [
              "parity_fit",
              "single_strike_parity",
              "spot"
            ],
            "title": "Forward Method",
            "type": "string"
          },
          "moment": {
            "title": "Moment",
            "type": "string"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ChainRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "spot": {
            "title": "Spot",
            "type": "number"
          }
        },
        "required": [
          "moment",
          "spot",
          "expiry",
          "forward",
          "discount",
          "forward_method",
          "rows"
        ],
        "title": "ChainResponse",
        "type": "object"
      },
...
      "Curve": {
        "description": "The chart's line: P&L at Expiry across a range of Spot values.\n\nTwo parallel arrays, as the prototype in #9 published them and as a chart consumes\nthem. Both lines a trader sees are P&L rather than Payoff (CONTEXT.md) - Payoff is\npremium-blind and is one subtraction away.",
        "properties": {
          "pnl_at_expiry": {
            "items": {
              "type": "number"
            },
            "title": "Pnl At Expiry",
            "type": "array"
          },
          "spot": {
            "items": {
              "type": "number"
            },
            "title": "Spot",
            "type": "array"
          }
        },
        "required": [
          "spot",
          "pnl_at_expiry"
        ],
        "title": "Curve",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PresetResponse": {
        "description": "What the picker offers, and what one of them builds.\n\nThe Legs come back as **requests**, not as resolved Legs: a Preset hands the client\nexactly what a trader would have picked by hand, and it goes back through the same\nendpoint. That is what makes \"analysing a Preset\" and \"selecting its Legs by hand\"\nthe same operation rather than two paths that agree.",
        "properties": {
          "legs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/LegRequest"
            },
            "title": "Legs",
            "type": "array"
          },
          "presets": {
            "items": {
              "type": "string"
            },
            "title": "Presets",
            "type": "array"
          }
        },
        "required": [
          "presets"
        ],
        "title": "PresetResponse",
        "type": "object"
      },
      "SessionResponse": {
        "description": "What a client needs before it can ask for anything else.\n\nEvery other endpoint takes a `moment`, and until this existed there was no way to\nlearn which moments there are. The frontend held the list in a generated fixture\ninstead, which made the set of tradeable minutes a property of a build script rather\nthan of the data - and free to drift from it silently.\n\nBounds and counts as well as the list itself, because a slider needs its ends and a\ntrader reading \"376\" learns something the array does not tell them at a glance.\n\nA session is **one date and one Expiry** (#68), and it carries the two lists a client\npicks the next one from. That is what makes the dropdowns above the Chain open\nwithout reading a data file, and it is the same argument as the paragraph above: a\nlist of dates written by a build script is a list that can drift from the tree the\nengine will actually serve.\n\n`date` and `expiry` name the pair this response describes, and they are **not\nnecessarily the pair that was asked for**. A request naming a date and an Expiry that\ndate did not trade is resolved to one that exists rather than refused, so a client\nrenders these two fields rather than what it sent.",
        "properties": {
          "date": {
            "title": "Date",
            "type": "string"
          },
          "dates": {
            "items": {
              "type": "string"
            },
            "title": "Dates",
            "type": "array"
          },
          "expiries": {
            "items": {
              "type": "string"
            },
            "title": "Expiries",
            "type": "array"
          },
          "expiry": {
            "title": "Expiry",
            "type": "string"
          },
          "first_moment": {
            "title": "First Moment",
            "type": "string"
          },
          "last_moment": {
            "title": "Last Moment",
            "type": "string"
          },
          "moment_count": {
            "minimum": 1.0,
            "title": "Moment Count",
            "type": "integer"
          },
          "moments": {
            "items": {
              "type": "string"
            },
            "title": "Moments",
            "type": "array"
          },
          "presets": {
            "items": {
              "type": "string"
            },
            "title": "Presets",
            "type": "array"
          },
          "strike_max": {
            "title": "Strike Max",
            "type": "number"
          },
          "strike_min": {
            "title": "Strike Min",
            "type": "number"
          }
        },
        "required": [
          "date",
          "dates",
          "moments",
          "moment_count",
          "first_moment",
          "last_moment",
          "expiry",
          "expiries",
          "strike_min",
          "strike_max",
          "presets"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
      "SummaryResponse": {
        "description": "The header, at one minute - and nothing about any strike (#69).\n\nSpot, the Forward, the Discount Factor and the at-the-money volatility are facts about\nthe **minute**. In the stored Chain they repeat across every strike of that minute, so\na header that read them there opened the largest artifact in the tree to take four\nnumbers out of about 196 identical copies of each. They are stored once instead, one\nrow a minute, and this is that row on the wire.\n\nDragging the time control moves the header 375 times across a session. Every one of\nthose was a read of the Chain and is now a lookup.\n\n**Every figure here is the one `/chain` publishes for the same minute.** That is not a\ncoincidence to be checked occasionally, it is the reason the split is safe: the build\nreduces the Chain frame it is about to write, so there is no second derivation that\ncould drift. Two artifacts describing one minute that can disagree would be worse than\none artifact.",
        "properties": {
          "atm_iv": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atm Iv"
          },
          "atm_strike": {
            "title": "Atm Strike",
            "type": "number"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "discount": {
            "title": "Discount",
            "type": "number"
          },
          "expiry": {
            "title": "Expiry",
            "type": "string"
          },
          "forward": {
            "title": "Forward",
            "type": "number"
          },
          "forward_method": {
            "enum": [
              "parity_fit",
              "single_strike_parity",
              "spot"
            ],
            "title": "Forward Method",
            "type": "string"
          },
          "moment": {
            "title": "Moment",
            "type": "string"
          },
          "spot": {
            "title": "Spot",
            "type": "number"
          }
        },
        "required": [
          "moment",
          "date",
          "expiry",
          "spot",
          "forward",
          "discount",
          "forward_method",
          "atm_strike"
        ],
        "title": "SummaryResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "convex-hedge payoff engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/analyse": {
      "post": {
        "description": "Everything about one Strategy, as-of one moment, in one response.\n\nThe Expiry comes off the **Legs** (#71) and the request no longer carries one. It is\nsettled before anything is read, because it is what the reads are keyed by: the header\nfigures below belong to one series at one minute, and a Strategy that named two would\notherwise take them from whichever the store listed first.",
        "operationId": "analyse_analyse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyse"
      }
    },
    "/chain": {
      "get": {
        "description": "The Chain as-of a moment: one row per strike, call and put either side.\n\n`date` is the trading date, which is what the store is keyed by and what a link\ncarries now that a control picks one (#68). Omitted, it is taken off the moment: the\nsession runs 03:45 to 10:00 UTC, so it never crosses a midnight and the two cannot\ndisagree. Naming it is what makes the twenty-three days #67 built reachable without a\nclient having to know that.\n\n`expiry` is the other partition key, spelled as the response spells it - `10FEB26`.\nOmitted, the Chain covers every series that traded that day, which is what it meant\nbefore there was a second dropdown to name one. Named, it covers exactly one; the day\na second series exists, the difference is a Chain against two interleaved, and\nnothing on screen would say which was being read.\n\n**Strict, unlike `/session`.** A pair the store does not hold is a 404 naming what it\ndoes hold, not a quiet substitution: a Chain that is not the one that was asked for\nis indistinguishable from one that is.",
        "operationId": "read_chain_chain_get",
        "parameters": [
          {
            "in": "query",
            "name": "moment",
            "required": true,
            "schema": {
              "title": "Moment",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "expiry",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expiry"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Chain"
      }
    },
...
    "/presets/{name}": {
      "get": {
        "description": "The Legs a Preset builds, as a trader would have picked them by hand.\n\nReturned as requests rather than analysed here, so that choosing a Preset and\npicking its Legs off the Chain are the same operation and not two paths that agree.\n\n`date` and `expiry` name the series the Legs are built in, spelled as `/chain` spells\nthem, because a Leg carries its own Expiry now (#71) and a Preset is a shape rather\nthan a set of contracts. Omitted, they are the Chain's own answer for this minute -\nnot a constant, which would hand a trader Legs in a series they are not looking at.",
        "operationId": "build_preset_presets__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "moment",
            "required": true,
            "schema": {
              "title": "Moment",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "expiry",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expiry"
            }
          },
          {
            "in": "query",
            "name": "strike",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Strike"
            }
          },
          {
            "in": "query",
            "name": "width",
            "required": false,
            "schema": {
              "default": 200.0,
              "title": "Width",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Direction",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Preset"
      }
    },
    "/session": {
      "get": {
        "description": "One day and one Expiry: which minutes exist, what else exists, what the picker offers.\n\nAsked whenever either dropdown moves. Everything else takes a `moment`, and this is\nthe only way to learn which moments there are - so it is served from the data rather\nthan from a clock, and a client that renders 376 stops is rendering 376 minutes that\nquoted. `dates` and `expiries` are the same argument one level up (#68): the two\ndropdowns are populated from here and never from a file, because a generated list is\nfree to describe a day the engine would not serve.\n\n**The pair is resolved, not merely validated.** A trader changes the date and the\nExpiry in the URL is one interaction behind; `catalog.resolve` returns a pair the\nstore actually holds, and `date` and `expiry` in the response say which - so what\ncomes back is never an empty Chain and never a session describing a day `/chain`\nwould refuse. `/chain` itself is strict, because it is asked for one specific thing.",
        "operationId": "read_session_session_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "expiry",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expiry"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Session"
      }
    },
    "/summary": {
      "get": {
        "description": "The header at one minute: Spot, the Forward, the Discount Factor, the money (#69).\n\n**This is what moving the time control asks for.** The four figures belong to the\nminute rather than to a strike, and in the stored Chain they repeat across every one\nof that minute's ~196 rows; a drag across a session asked for them 375 times and\nopened the million-row artifact 375 times to do it. They live in a 375-row-a-day\nSummary now, and this reads one row of it. Nothing under this endpoint opens the Chain.\n\nSame keys and same rules as `/chain`: `date` off the moment when it is omitted,\n`expiry` spelled `10FEB26`, and **strict** - a pair the store does not hold is a 404\nnaming what it does, never a quiet substitution.\n\nThe numbers are the ones `/chain` publishes for the same minute, and by construction\nrather than by coincidence: the build reduces the Chain frame it is about to write.",
        "operationId": "read_summary_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "moment",
            "required": true,
            "schema": {
              "title": "Moment",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "expiry",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expiry"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Summary"
      }
    }
  }
}